+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format of this changelog is based on
 
   - Fixed incorrect conversion and reflection of split BSplines
   - Added FAQ entry about MeshSized/OptionalEntity styling on Paths
+  - Fixed issue causing duplicate `Cell` names with paths and composite components, where rendering would use the component's name rather than a unique name
 
 ## 1.4.2 (2025-07-16)